Subtitle: Adjust the position of stop parser and break. [1/1]
PD#SWPL-75238

Problem:
Serial input setprop persist.vendor.tif.subtitleflg
0xFF,DUT reboot,CC is not displayed

Solution:
The cc parser is created by default, and the
waiting time is 100ms. At this time, the userdata
has a high probability that the cc data is not ready,
which will cause the parser object to be deleted.

Verify:
local

Signed-off-by: shuai.zhang <shuai.zhang@amlogic.com>
Change-Id: I35d562910aa787c79292e4af42cb852e21165176